refactor: Moved drawable object files into subfolder
