+++ src/drawables/IDrawableObject.ts
@@ -1,0 +1,20 @@
+import { Group, Vector3, Euler } from "three";
+
+export interface IDrawableObject {
+  cleanup(): void;
+  setVisible(visible: boolean): void;
+  doRender(): void;
+  get3dObject(): Group;
+  setTranslation(translation: Vector3): void;
+  setScale(scale: Vector3): void;
+  /**
+   * Optional. Should be called when parent transforms are updated.
+   */
+  onParentTransformUpdated?(): void;
+  setRotation(eulerXYZ: Euler): void;
+  setFlipAxes(flipX: number, flipY: number, flipZ: number): void;
+  setOrthoThickness(thickness: number): void;
+  setResolution(x: number, y: number): void;
+  setAxisClip(axis: "x" | "y" | "z", minval: number, maxval: number, _isOrthoAxis: boolean): void;
+  updateClipRegion(xmin: number, xmax: number, ymin: number, ymax: number, zmin: number, zmax: number): void;
+}
