+++ src/drawables/lines/Line3d.ts
@@ -2,9 +2,9 @@ import { Color, DataTexture, FloatType, LinearFilter, RGBAFormat, Texture } from
 import { LineSegments2 } from "three/addons/lines/LineSegments2.js";
 import { LineSegmentsGeometry } from "three/addons/lines/LineSegmentsGeometry.js";
 
-import { IDrawableObject } from "./types.js";
-import { MESH_NO_PICK_OCCLUSION_LAYER, OVERLAY_LAYER } from "./ThreeJsPanel.js";
-import BaseDrawableMeshObject from "./BaseDrawableMeshObject.js";
+import { IDrawableObject } from "../IDrawableObject.js";
+import { MESH_NO_PICK_OCCLUSION_LAYER, OVERLAY_LAYER } from "../../ThreeJsPanel.js";
+import BaseDrawableMeshObject from "../BaseDrawableMeshObject.js";
 import SubrangeLineMaterial from "./SubrangeLineMaterial.js";
 
 const DEFAULT_VERTEX_BUFFER_SIZE = 1020;
